elec module form changes
