+++ BrainQuake/gui_forms/elec_form.py
@@ -1,4 +1,4 @@
-#! /usr/bin/python3.6
+#! /usr/bin/python3.7
 # -- coding: utf-8 -- **
 
 from PyQt5 import QtCore, QtGui, QtWidgets
